feat(start): start voca-test-front

+++ src/App.js
@@ -8,11 +8,15 @@ import './App.css';
 
 function App(props) {
 
-  let url = 'http://localhost:5000/login';
+  let url = 'http://3.34.140.114:5000/login';
 
   const [login, setLogin] = useState("실패");
 
   const [loginData, setLoginData] = useState({ code : '', password : ''});
+
+  const [code, clearCode] = useState('');
+
+  const [password, clearPassword] = useState('');
 
   const onChange = (e) => {
     const { value, name } = e.target; 
@@ -26,26 +30,45 @@ function App(props) {
     setLogin('성공');
   }
 
-  const refresh = () => {
-    window.location.reload();
-  }
-
   const goLogin = async () => {
     await axios.get(url);  
     setLogin('실패');
   }
 
+  const feedback = () => {
+    if (window.confirm('피드백 하러 가시겠습니까?')) {
+      return window.location.href = "https://docs.google.com/forms/d/e/1FAIpQLSe5f-w3Lw-m2J-WPaKTbdAveReoWglcXu53QhFCCYEFLAY87w/viewform?usp=sf_link"
+    } else {
+      console.log('false');
+      return;
+    }
+  }
+
   const check = (e) => {
     e.preventDefault();
-    axios.post(url, loginData)
-      .then(req => {
-        const token = req.data.userId;
-        localStorage.setItem('userId', token);
-        goMain();
-      })
-      .catch(err => {
-        window.alert('아이디, 비밀번호를 확인해주세요.');
-      })
+    if (loginData.code === '') {
+      window.alert('단어장을 입력해주세요.');
+      return;
+    } else {
+      if (loginData.password === '') {
+        window.alert('비밀번호를 입력해주세요.');
+        return;
+      } else {
+        axios.post(url, loginData)
+          .then(req => {
+            setLoginData({ code : '', password : ''});
+            clearCode('');
+            clearPassword('');
+            const token = req.data.userId;
+            localStorage.setItem('userId', token);
+            console.log(req);
+            goMain();
+          })
+          .catch(err => {
+            window.alert('아이디, 비밀번호를 확인해주세요.');
+          })
+      }
+    }
   }
 
   useEffect(() => {
@@ -65,7 +88,14 @@ function App(props) {
   return (
     <div style={{ height : "100%"}}>
       <div id="nav">
-        <button id="logo" onClick={refresh}>
+        <button 
+          id="logo"
+          onClick={feedback}
+          style={{
+            textDecoration: 'none',
+            color: '#FAF3DD',
+          }}
+        >
           <div>Word Note</div>
           <div 
             style={{ 
@@ -85,16 +115,25 @@ function App(props) {
             onSubmit={check}
           >
             <input 
-              onChange={(e) => onChange(e)}
+              onChange={(e) => {
+                onChange(e);
+                clearCode(e.target.value);
+              }}
               className="login-inputs"
               placeholder="단어장 입력"
               name="code"
               type='text'
+              value={code}
+              style={{imeMode : 'active'}}
             />
             <input 
-              onChange={(e) => onChange(e)}
+              onChange={(e) => {
+                onChange(e);
+                clearPassword(e.target.value);
+              }}
               className="login-inputs"
               placeholder="비밀번호 입력"
+              value={password}
               name="password"
               type='password'
             />
